requirements.txt
torch>=2.2
torchvision>=0.17
hydra-core>=1.3
omegaconf>=2.3
albumentations>=1.4
opencv-python>=4.9
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
timm>=1.0
scikit-learn>=1.3
